third_party/blink/web_tests/flag-specific/highdpi/accessibility/image-map-bounds-expected.txt

This is a testharness.js-based test.
FAIL Test bounds of image map links assert_equals: expected 210 but got 315
Harness: the test ran to completion.
